pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "graph_transformer"
version = "0.1.0"
description = "Knowledge graph based transformer"
requires-python = ">=3.9"
dependencies = [
  "pandas==1.4.*",
  "numpy==1.21.*",
  "scikit-learn==1.1.*",
  "tqdm",
  "python-dotenv",
  "dill",
]

[project.optional-dependencies]
dev = [
  "isort==5.10.1",
  "black==22.3.0",
  "mypy==0.991",
  "pytest==7.*",
  "pytest-cov",
  "pre-commit==2.20.*",
  "jupyterlab==3.4.*",
]

[project.scripts]
run_graph_transformer = "graph_transformer.cli.run_graph_transformer:cli_main"

[tool.setuptools.packages.find]
where = ["src"]
# include = ["pkg*"]  # alternatively: `exclude = ["additional*"]`
include = ["graph_transformer"]
namespaces = false

[tool.black]
line-length = 79
target-version = ['py39']
include = '''
(
    ^/tests/.*\.pyi?$
  | ^/graph_transformer/.*\.pyi?$
  | ^/snippets/.*\.pyi?$
  | ^/setup[.]py
  | ^/conftest[.]py
)
'''
exclude = '''
(  \.eggs
  | \.git
  | \.venv
  | build
  | dist
  | venv
  | __pycache__
)
'''

[tool.pycln]
include = '''
(
    ^/tests/
  | ^/graph_transformer/
  | ^/snippets/
)
'''
all = true

[tool.isort]
profile = "black"
src_paths = ["graph_transformer", "tests", "snippets" ]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 79


[tool.mypy]
files = '''
(
    ^/graph_transformer/.*\.pyi?$
    | ^/tests/.*\.pyi?$
    | ^/snippets/.*\.pyi?$
)
'''
python_version = 3.9
[[tool.mypy.overrides]]
module = 'pytest'
ignore_missing_imports = true

[tool.coverage.run]
source = ['graph_transformer']
omit = ['graph_transformer/migrations/*',
  ]
